app1/requirements.txt
django==4.2.11
djangorestframework==3.15.1
django-allauth==0.61.1
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==22.0.0
